refatorando backend

+++ src/main/java/com/union/model/repository/FuncionarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.union.model.entity.Funcionario;
 
-public interface FuncionarioRepository extends JpaRepository<Funcionario, Long>{
+public interface FuncionarioRepository extends JpaRepository<Funcionario, Integer>{
 
 }
